init: First step to basic Linux userspace
Placeholder for init binary and shell.

Implement source for binary that can be run by Linux kernel as PID 1.
After being executed, it mounts procfs and sysfs then runs a
rudimentary shell to interact with the contents of the VFS (now assumed
to be in initramfs).

Available commands: 'reset', 'halt' and 'poweroff' to stop the kernel and
'cd', 'ls' and 'cat' to view directories and files.

See README.md for more details.

Signed-off-by: Alex Juncu <alexj@linux.com>